(PC-5884) Move LogoutButton from cheatcodes to auth/components
