+++ src/features/cheatcodes/pages/CheatCodes.tsx
@@ -6,9 +6,9 @@ import { Text, Alert, Button } from 'react-native'
 import styled from 'styled-components/native'
 
 import { api } from 'api/api'
+import { LogoutButton } from 'features/auth/components/LogoutButton'
 import { CodePushButton } from 'features/cheatcodes/components/CodePushButton'
 import { CrashTestButton } from 'features/cheatcodes/components/CrashTestButton'
-import { LogoutButton } from 'features/cheatcodes/components/LogoutButton'
 import { NavigateHomeButton } from 'features/cheatcodes/components/NavigateHomeButton/NavigateHomeButton'
 import { RootStackParamList } from 'features/navigation/RootNavigator'
 import { env } from 'libs/environment'
